+++ jubatus/core/classifier/passive_aggressive_2.cpp
@@ -50,15 +50,15 @@ void passive_aggressive_2::train(const common::sfv_t& sfv,
   labels_.get_model()->increment(label);
 
   string incorrect_label;
-  float margin = calc_margin(sfv, label, incorrect_label);
-  float loss = 1.f + margin;
+  double margin = calc_margin(sfv, label, incorrect_label);
+  double loss = 1.0 + margin;
 
-  if (loss < 0.f) {
+  if (loss < 0.0) {
     storage_->register_label(label);
     return;
   }
-  float sfv_norm = squared_norm(sfv);
-  if (sfv_norm == 0.f) {
+  double sfv_norm = squared_norm(sfv);
+  if (sfv_norm == 0.0) {
     storage_->register_label(label);
     return;
   }
