+++ jubatus/core/fv_converter/dynamic_combination_feature.hpp
@@ -41,7 +41,7 @@ class dynamic_combination_feature
       const std::string& key,
       double value_left,
       double value_right,
-      std::vector<std::pair<std::string, float> >& ret_fv) const;
+      std::vector<std::pair<std::string, double> >& ret_fv) const;
 
   /**
    * Return true if the add_feature function is commutative with any
